+++ src/components/home/ScrollTitleWrapper.js
@@ -3,10 +3,10 @@ import PropTypes from "prop-types"
 import styled from "styled-components"
 import VisibilitySensor from "react-visibility-sensor"
 
-import TypingDisplay from "../components/TypingDisplay"
+import TypingDisplay from "../common/TypingDisplay"
 
-import { colors, sizes, devices } from "../constants/home"
-import { toDualColors } from "../helpers"
+import { colors, sizes, devices } from "../../constants/home"
+import { toDualColors } from "../../helpers"
 
 const Wrapper = styled.section`
   width: 100%;
